Extract base mod select overlay panel presentation logic
